Rejig for Component Ridiculousness
Reworked most scripts to (I think) fit Unity's idea of a component
structure, with horrible classes which don't actually represent objects
but things like "ExplodeOnImpact". My need to conform to OOP hurts so,
so much, but apparently this is the future so who am I to argue.  It
seems to me like this component lark is too eager to expose things
publicly, removes the "thinginess" from "things", and makes me often
wonder "where the hell should I put that function if there isn't a
"Planet" class because my "Planet"s are really just assorted scripts
tacked together. I don't see any particularly strong benefits other than
making it easier for designers to create a monkey that has the head of a
frog and the wings of a fox.